borrado en cascada de ajustes, ajustes web y turno si se borra una tienda cone stos y borrado de certificado afip si ya existe uno

+++ PointOfSale.Model/Ajustes.cs
@@ -15,7 +15,7 @@ namespace PointOfSale.Model
         public bool? ControlStock { get; set; }
         public string? NombreTiendaTicket { get; set; }
         public string? NombreImpresora { get; set; }
-        public int? MinimoIdentificarConsumidor { get; set; }
+        public long? MinimoIdentificarConsumidor { get; set; }
 
         public int IdTienda { get; set; }
         public Tienda? Tienda { get; set; }
