+++ PointOfSale/Models/VMAjustes.cs
@@ -7,7 +7,7 @@
         public bool? ImprimirDefault { get; set; }
         public bool? ControlStock { get; set; }
         public string? NombreImpresora { get; set; }
-        public int? MinimoIdentificarConsumidor { get; set; }
+        public long? MinimoIdentificarConsumidor { get; set; }
         public string? NombreTiendaTicket { get; set; }
         public int IdTienda { get; set; }
